refactor: types and interfaces for DTO
