+++ src/routes/user/dto/create-user.dto.ts
@@ -1,0 +1,11 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class CreateUserDto {
+  @IsNotEmpty()
+  @IsString()
+  readonly login: string;
+
+  @IsNotEmpty()
+  @IsString()
+  readonly password: string;
+}
