+++ src/routes/user/dto/update-user.dto.ts
@@ -1,15 +1,5 @@
 import { IsNotEmpty, IsString } from 'class-validator';
 
-export class CreateUserDto {
-  @IsNotEmpty()
-  @IsString()
-  readonly login: string;
-
-  @IsNotEmpty()
-  @IsString()
-  readonly password: string;
-}
-
 export class UpdatePasswordDto {
   @IsNotEmpty()
   @IsString()
