+++ src/types.ts
@@ -1,4 +1,4 @@
-export interface User {
+export interface IUser {
   id: string; // uuid v4
   login: string;
   password: string;
